+++ levelEditor/main.js
@@ -287,7 +287,7 @@ const blockProperty = {
   33: ["ID"],
   34: ["ID"],
   35: ["ID"],
-  41: ["X", "Y", "Absolute"],
+  41: ["X", "Y", "Absolute", "No Camera Easing"],
   46: ["Text"],
   47: ["Power"],
   48: ["Gravity", "Horizontal"],
@@ -350,7 +350,7 @@ const defaultProperty = {
   33: [0],
   34: [0],
   35: [0],
-  41: [0, 0, false],
+  41: [0, 0, false, false],
   46: ["Text"],
   47: [275],
   48: [325, false],
@@ -410,7 +410,7 @@ const propertyType = {
   33: ["integer"],
   34: ["integer"],
   35: ["integer"],
-  41: ["number", "number", "boolean"],
+  41: ["number", "number", "boolean", "boolean"],
   46: ["any"],
   47: ["number"],
   48: ["number", "boolean"],
@@ -479,7 +479,7 @@ const propertyLimit = {
   33: [[0, 99]],
   34: [[0, 99]],
   35: [[0, 99]],
-  41: ["none", "none", "none"],
+  41: ["none", "none", "none", "none"],
   46: ["none"],
   47: [[0, 1000]],
   48: [[-2000, 2000], "none"],
@@ -1476,6 +1476,7 @@ function nextFrame(timeStamp) {
                         (y + props[2]) * blockSize +
                         (blockSize - player.size) / 2;
                     }
+                    if (props[4]) adjustScreen(true);
                     break;
                   // coin
                   case 77: