+++ changelog.md
@@ -1277,3 +1277,9 @@ Editor
 - Refactored and optimized code for camera movement and grid display
   - This might've also fixed some visual glitches with grid display when changing browser zoom
 - Fixed a bug where cancelling an "increase level size by specified amount" prompt caused it to display "Invalid input!"
+
+### 21 Jun. 2025
+
+Editor
+
+- Added the 'No Camera Easing' property to Portals, which, if true, prevents the camera from easing when you teleport using a Portal.